requirements.txt
Django==4.0.5
gunicorn==20.1.0
Pillow==9.1.1
python-dotenv==0.20.0
whitenoise==6.2.0
django-debug-toolbar==3.5.0
django-admin-interface==0.19.1
django-colorfield==0.7.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
markdown==3.4.1
djangorestframework==3.13.1
